requirements.txt
flake8==5.0.4
load-dotenv==0.1.0
mccabe==0.7.0
numpy==1.23.4
paho-mqtt==1.6.1
pandas==1.5.1
pika==1.3.1
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
tabulate==0.9.0
